aeneas_pipeline/edge-modules/AeneasEdgeSolution/modules/FileSenderModule/requirements.txt
azure-iot-device~=2.7.0
azure-storage-blob
